requirements/local.txt
-r base.txt
-e .
-e ../base
-e ../login
-e ../mail
factory-boy
psycopg2
pytest-coverage
pytest-django
pytest-flakes
pytest-pep8
